Major refactoring, and restructing of student view tab
Addresses #33
