+++ packages/portal/frontend/src/app/views/GradingPage.ts
@@ -1,19 +1,19 @@
 import {OnsInputElement, OnsSearchInputElement} from "onsenui";
-import Log from "../../../../../../common/Log";
+import Log from "../../../../../common/Log";
 import {
     AssignmentGrade,
     AssignmentRubric,
     QuestionGrade, SubQuestionGrade,
     SubQuestionRubric
-} from "../../../../../../common/types/CS340Types";
-import {DeliverableTransport, RepositoryTransport, TeamTransport} from "../../../../../../common/types/PortalTypes";
-import {Factory} from "../../Factory";
-import {UI} from "../../util/UI";
-import {AdminDeliverablesTab} from "../AdminDeliverablesTab";
-import {AdminPage} from "../AdminPage";
-import {AdminResultsTab} from "../AdminResultsTab";
-import {AdminView} from "../AdminView";
+} from "../../../../../common/types/CS340Types";
+import {DeliverableTransport, RepositoryTransport, TeamTransport} from "../../../../../common/types/PortalTypes";
+import {Factory} from "../Factory";
+import {UI} from "../util/UI";
+import {AdminDeliverablesTab} from "./AdminDeliverablesTab";
 import {AdminMarkingTab} from "./AdminMarkingTab";
+import {AdminPage} from "./AdminPage";
+import {AdminResultsTab} from "./AdminResultsTab";
+import {AdminView} from "./AdminView";
 
 const ERROR_POTENTIAL_INCORRECT_INPUT: string = "input triggered warning";
 const ERROR_INVALID_INPUT: string = "invalid input";
@@ -24,7 +24,17 @@ const WARN_EMPTY_FIELD: string = "empty field";
 
 export class GradingPageView extends AdminPage {
 
-    private UBC_LETTER_GRADES: Map<string, {lower: number, upper: number}> = new Map<string, {lower: number, upper: number}>();
+    public static UBC_LETTER_GRADES: Map<string, {lower: number, upper: number}> = new Map<string, {lower: number, upper: number}>([
+        ["A+", {lower: 90  , upper: 100}],
+        ["A",  {lower: 85  , upper: 89}],
+        ["A-", {lower: 80  , upper: 84}],
+        ["B+", {lower: 76  , upper: 79}],
+        ["B",  {lower: 72  , upper: 75}],
+        ["B-", {lower: 68  , upper: 71}],
+        ["C+", {lower: 64  , upper: 67}],
+        ["C",  {lower: 60  , upper: 63}],
+        ["F",  {lower: 0   , upper: 59}],
+    ]);
     // private students: string[];
     // private isTeam: boolean;
     // private deliverableId: string;
@@ -41,15 +51,15 @@ export class GradingPageView extends AdminPage {
     }
 
     public async init(opts: any): Promise<void> {
-        this.UBC_LETTER_GRADES.set("A+",    {lower: 90  , upper: 100});
-        this.UBC_LETTER_GRADES.set("A",     {lower: 85  , upper: 89});
-        this.UBC_LETTER_GRADES.set("A-",    {lower: 80  , upper: 84});
-        this.UBC_LETTER_GRADES.set("B+",    {lower: 76  , upper: 79});
-        this.UBC_LETTER_GRADES.set("B",     {lower: 72  , upper: 75});
-        this.UBC_LETTER_GRADES.set("B-",    {lower: 68  , upper: 71});
-        this.UBC_LETTER_GRADES.set("C+",    {lower: 64  , upper: 67});
-        this.UBC_LETTER_GRADES.set("C",     {lower: 60  , upper: 63});
-        this.UBC_LETTER_GRADES.set("F",     {lower: 0   , upper: 59});
+        // GradingPageView.UBC_LETTER_GRADES.set("A+",    {lower: 90  , upper: 100});
+        // GradingPageView.UBC_LETTER_GRADES.set("A",     {lower: 85  , upper: 89});
+        // GradingPageView.UBC_LETTER_GRADES.set("A-",    {lower: 80  , upper: 84});
+        // GradingPageView.UBC_LETTER_GRADES.set("B+",    {lower: 76  , upper: 79});
+        // GradingPageView.UBC_LETTER_GRADES.set("B",     {lower: 72  , upper: 75});
+        // GradingPageView.UBC_LETTER_GRADES.set("B-",    {lower: 68  , upper: 71});
+        // GradingPageView.UBC_LETTER_GRADES.set("C+",    {lower: 64  , upper: 67});
+        // GradingPageView.UBC_LETTER_GRADES.set("C",     {lower: 60  , upper: 63});
+        // GradingPageView.UBC_LETTER_GRADES.set("F",     {lower: 0   , upper: 59});
         Log.info(`GradingPage::init(..) - opts: ${JSON.stringify(opts)}`);
         this.studentId = opts.sid;
         this.assignmentId = opts.aid;
@@ -195,7 +205,7 @@ export class GradingPageView extends AdminPage {
                 if (previousSubmission === null || !previousSubmission.questions[i].subQuestions[j].graded) {
                     gradeInputElement.setAttribute("placeHolder", subQuestion.name);
                 } else {
-                    const letterGrade: string = this.getLetterGrade(
+                    const letterGrade: string = GradingPageView.getLetterGrade(
                         previousSubmission.questions[i].subQuestions[j].grade,
                         subQuestion.outOf);
                     gradeInputElement.setAttribute("placeHolder", letterGrade);
@@ -407,7 +417,7 @@ export class GradingPageView extends AdminPage {
                     errorStatus = true;
                 } else {
                     const gradeLetter: string = gradeInputElement.value.toUpperCase();
-                    if (!this.UBC_LETTER_GRADES.has(gradeLetter)) {
+                    if (!GradingPageView.UBC_LETTER_GRADES.has(gradeLetter)) {
                         gradeValue = 0;
                         if (!warnStatus) {
                             warnComment = WARN_EMPTY_FIELD;
@@ -416,7 +426,7 @@ export class GradingPageView extends AdminPage {
                         graded = false;
                         errorElement.innerHTML = "Warning: Input field is empty";
                     } else {
-                        const gradeRange = this.UBC_LETTER_GRADES.get(gradeLetter);
+                        const gradeRange = GradingPageView.UBC_LETTER_GRADES.get(gradeLetter);
                         const multiplier = Math.ceil((gradeRange.upper + gradeRange.lower) / 2) / 100;
                         const outOf = parseFloat(gradeInputElement.getAttribute("data-outOf"));
                         gradeValue = multiplier * outOf;
@@ -499,7 +509,7 @@ export class GradingPageView extends AdminPage {
             const teamResponse = await fetch(teamURL, teamOptions);
             if (teamResponse.status !== 200) {
                 const errJson = await teamResponse.json();
-                Log.error("MDSAdminView::submitGrade(..) - Error: " + errJson.error);
+                Log.error("CS340AdminView::submitGrade(..) - Error: " + errJson.error);
                 UI.notification("Unable to save grade to team; unable to find correct team");
                 return null;
             } else {
@@ -517,7 +527,7 @@ export class GradingPageView extends AdminPage {
     }
 
     public async submitGradeRecord(aid: string, personIds: string[], questionArray: QuestionGrade[]): Promise<boolean> {
-        Log.info("MDSAdminView::submitGradeRecord(..) - start");
+        Log.info("CS340AdminView::submitGradeRecord(..) - start");
         const allPromises: Array<Promise<any>> = [];
         UI.showModal("Submitting grade(s), please wait...");
 
@@ -552,7 +562,7 @@ export class GradingPageView extends AdminPage {
             Log.info("GradingPage::submitGrade() - response from api " + response);
             if (response.status !== 200) {
                 const errResponse = await response.json();
-                Log.trace("MDSAdminView::submitGrade() - error submitting grades, code: " +
+                Log.trace("CS340AdminView::submitGrade() - error submitting grades, code: " +
                     response.status + " error: " + response.statusText);
                 // alert(errResponse.error);
                 UI.showAlert(errResponse.error);
@@ -561,7 +571,7 @@ export class GradingPageView extends AdminPage {
             }
         }
         UI.hideModal();
-        Log.info("MDSAdminView::submitGradeRecord(..) - end");
+        Log.info("CS340AdminView::submitGradeRecord(..) - end");
 
         return true;
     }
@@ -621,15 +631,24 @@ export class GradingPageView extends AdminPage {
     /**
      *
      */
-    private getLetterGrade(grade: number, outOf: number): string {
+    public static getLetterGrade(grade: number, outOf: number): string {
         const gradePercent = (grade / outOf) * 100;
-        for (const key of this.UBC_LETTER_GRADES.keys()) {
-            const range = this.UBC_LETTER_GRADES.get(key);
+        for (const key of GradingPageView.UBC_LETTER_GRADES.keys()) {
+            const range = GradingPageView.UBC_LETTER_GRADES.get(key);
             if (range.lower <= gradePercent && gradePercent <= range.upper) {
                 return key;
             }
         }
 
         return "";
+    }
+
+    public static getLetterGradeMidpoint(letterGrade: string): number {
+        const gradeRange = GradingPageView.UBC_LETTER_GRADES.get(letterGrade);
+        if (gradeRange === null) {
+            return 0;
+        }
+
+        return (gradeRange.upper + gradeRange.lower) / 2;
     }
 }
